+++ test/libsolidity/syntaxTests/abiEncoder/conflicting_settings_reverse.sol
@@ -1,0 +1,4 @@
+pragma abicoder v1;
+pragma experimental ABIEncoderV2;
+// ----
+// SyntaxError 8273: (20-53): ABI coder v1 has already been selected through "pragma abicoder v1".