+++ test/libsolidity/syntaxTests/abiEncoder/invalid_pragma_value.sol
@@ -1,0 +1,3 @@
+pragma abicoder something;
+// ----
+// SyntaxError 2745: (0-26): Expected either "pragma abicoder v1" or "pragma abicoder v2".